Deleted old files:  - fe-notes  - db-schema  - types.tsx Rename:  - global.scss > App.scss  - interfaces.tsx > interfaces.ts Workspace: add organize imports on save
