--- DB schema draft.txt
@@ -1,27 +1,0 @@
-I need the PostGIS postgre extension to handle spatial data.
-
-type USER_ROLE = enum { urban_planner, resident, developer }
-type LINK_TYPE = enum { direct, collateral, projection, update}
-//TODO: introduce DOCUMENT.links jsonb validation based on enum USER_ROLE
-
-DOCUMENT {
-    id: serial,
-    title: varchar(128),
-    description: text,
-    coordinates: geography(POINT), // BEWARE PostGIS expects lng then lat 
-    links: jsonb schema { targetDocumentId: link_type[]} example { '3': ["update", "collateral"], '1': ["projection"]}
-    attachments: int[] (foreign key in ATTACHMENT)
-}
-
-ATTACHMENT {
-    id: serial,
-    url: text
-}
-
-USER {
-    email: varchar(255),
-    name: varchar(64),
-    surname: varchar(64),
-    password_hash: varchar(60),
-    role: USER_ROLE
-}
