+++ client/src/main.tsx
@@ -1,8 +1,7 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import "../src/styles/global.scss";
+import App from "./App.tsx";
 
 const router = createBrowserRouter([
   {
