--- frontend-notes.md
@@ -1,35 +1,0 @@
-
-
-Main page 
--   Inserire navbar in alto
--   POSSIBILE IDEA: quando clicco su documento -> slide laterale da dx a sx con prime info su documento
-
-        Lateral view:
-        -   Icon
-        -   Title                       US1
-        -   Stakeholders                
-        -   Georeference                US3 -> update at US5
-        -   Scale
-        -   Issuance date
-        -   Type
-        -   Connections: vedi sotto     US2
-        -   Language ?
-        -   Pages ? 
-        -   Description (magari piu in alto)
-        -   Download Original document
-        -   Magari un button "more info" per far apparire i dettagli secondari
-  
-
-CONNECTIONS
-    CREAZIONE CONNECTIONS
--   Bottone "Create new connection" per aggiungere connections a un documento con "menu a tendina"
--   Button "Create new connection" e poi selezione direttamente dalla mappa o anche dal diagramma con doc in risalto
-    
-    VISUALIZZAZIONE CONNECTIONS
--   
-
-
-
-QUESTIONS
-- Se usiamo la barra laterale -> come lo implementiamo nella visualizzazione da smartphone?
-- Can the stakeholders be more than one?
